services/template/pyproject.toml
[tool.poetry]
name = "template-service"
version = "0.1.0"
description = "Service for managing and rendering multi-channel templates"
authors = ["Admin <admin@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
asyncpg = "^0.30.0"
alembic = "^1.12.1"
pydantic = {extras = ["email"], version = "^2.5.2"}
pydantic-settings = "^2.1.0"
jinja2 = "^3.1.2"
httpx = "^0.25.1"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
prometheus-fastapi-instrumentator = "^7.1.0"
opentelemetry-api = "^1.39.0"
opentelemetry-sdk = "^1.39.0"
opentelemetry-instrumentation-fastapi = "^0.60b0"
opentelemetry-exporter-otlp = "^1.39.0"
asgi-correlation-id = "^4.3.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
